bot: project structure
